% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_class.R
\name{clump_snps,GWAS-method}
\alias{clump_snps,GWAS-method}
\title{clump_snps}
\usage{
\S4method{clump_snps}{GWAS}(object, do_local = FALSE)
}
\arguments{
\item{object}{a GWAS object}

\item{do_local}{run on local machine, if false does on IEU servers}
}
\value{
a GWAS object with use_flag adjusted to only those SNPs that are left after data clumping
}
\description{
clump_snps
}
