% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_class.R
\name{clump_snps}
\alias{clump_snps}
\title{clump_snps}
\usage{
clump_snps(object, do_local = FALSE)
}
\arguments{
\item{object}{a GWAS object}

\item{do_local}{run on local machine, if false does on IEU servers}
}
\value{
a GWAS object with clumped data
}
\description{
clump_snps
}
