% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cis_gene_exposure.R
\name{create_cis_gene_exposure}
\alias{create_cis_gene_exposure}
\title{create_cis_gene_exposure}
\usage{
create_cis_gene_exposure(exposure_data, gene)
}
\arguments{
\item{exposure_data}{exposure data.table, or list of, as returned by TwoSampleMR::format_data}

\item{gene}{a Gene object or list of Gene objects}
}
\value{
a list of data.tables, one for each exposure:gene combination
}
\description{
create_cis_gene_exposure
}
