% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_significant_snps.R
\name{extract_significant_snps}
\alias{extract_significant_snps}
\title{extract_significant_snps}
\usage{
extract_significant_snps(
  filepath,
  pval,
  pval_col,
  snp_col,
  csv_out = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{filepath}{a file path to extract from}

\item{pval}{p-value filter (values less than this are kept)}

\item{pval_col}{the name of the column holding the p-value}

\item{snp_col}{the name of the column holding the snp id}

\item{csv_out}{a name to label the output .csv with, default is to use the filepath filename}

\item{overwrite}{whether or not to overwrite the cache .csv file}
}
\value{
the data table; and attribute file path to the filtered .csv file
}
\description{
extract_significant_snps
}
