% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_class.R
\name{extract_snps,GWAS-method}
\alias{extract_snps,GWAS-method}
\title{extract_snps}
\usage{
\S4method{extract_snps}{GWAS}(object)
}
\arguments{
\item{object}{a GWAS class object}
}
\value{
a valid GWAS class object with SNP extracted into the data slot based on provided
filtering parameters
}
\description{
extract_snps
}
