% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_class.R
\name{extract_snps}
\alias{extract_snps}
\title{extract_snps}
\usage{
extract_snps(object)
}
\arguments{
\item{object}{a GWAS objects}
}
\value{
GWAS object with SNPs extracted to data field
}
\description{
The generic function for extract_snps
}
