% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_class.R
\docType{class}
\name{Gene-class}
\alias{Gene-class}
\alias{Gene}
\title{Gene class}
\value{
a Gene object
}
\description{
Gene class
}
\section{Slots}{

\describe{
\item{\code{name}}{character. The name of the gene.}

\item{\code{assembly}}{character. The assembly information, e.g. GRCh37}

\item{\code{chromosome}}{numeric. The chromosome the gene is on.}

\item{\code{start}}{numeric. The start (number of bases).}

\item{\code{end}}{numeric. The end (number of bases).}

\item{\code{cis_tol}}{when defining SNPs on a gene, set a tolerance of how many bases either side of the gene}
}}

