% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_class.R
\name{genome_region<-,GWAS,GenomeRegion-method}
\alias{genome_region<-,GWAS,GenomeRegion-method}
\title{set genome_region}
\usage{
\S4method{genome_region}{GWAS,GenomeRegion}(object) <- value
}
\arguments{
\item{object}{a valid GWAS object}

\item{value}{a valid GenomeRegion object}
}
\value{
a valid GWAS object with GenomeRegion set
}
\description{
set genome_region
}
