% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_class.R
\name{genome_region<-}
\alias{genome_region<-}
\title{set genome_region generic function}
\usage{
genome_region(object) <- value
}
\arguments{
\item{object}{a valid GWAS object}

\item{value}{a valid GenomeRegion object}
}
\value{
a valid GWAS object with GenomeRegion set
}
\description{
set genome_region generic function
}
