% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_region_class.R
\docType{class}
\name{GenomeRegion-class}
\alias{GenomeRegion-class}
\alias{GenomeRegion}
\title{GenomeRegion class}
\value{
a GenomeRegion object
}
\description{
This class define an area of the genome and handles various parameters passed to the LD clumping
algorithms. This is useful when working with SNPs surrounding multiple gene targets of interest.
To perform normal genome wide SNP clumping simple create a GenomeRegion object with
'genome_wide=TRUE'.
}
\section{Slots}{

\describe{
\item{\code{name}}{character. The name of the gene or area of the genome referred to.}

\item{\code{assembly}}{character. The assembly information, e.g. GRCh37}

\item{\code{chromosome}}{numeric. The chromosome the gene is on.}

\item{\code{start}}{numeric. The start (number of bases).}

\item{\code{end}}{numeric. The end (number of bases).}

\item{\code{cis_tol}}{when defining SNPs on a gene, set a tolerance of how many bases either side of the
gene}

\item{\code{genome_wide}}{this is usually FALSE, if TRUE this is treated as a object placeholder and will
refer to the 'whole genome'}

\item{\code{clump_kb}}{Clumping window, default is 10000. (vector of length = length(exposure_data); or
1, in which case it will be recycled)}

\item{\code{clump_r2}}{Clumping r2 cutoff.  (vector of length = length(exposure_data); or 1, in which
case it will be recycled)}

\item{\code{clump_p1}}{Clumping sig level for index SNPs, default is 1.  (vector of length =
length(exposure_data); or 1, in which case it will be recycled)}

\item{\code{clump_p2}}{Clumping sig level for secondary SNPs, default is 1.  (vector of length =
length(exposure_data); or 1, in which case it will be recycled)}

\item{\code{pop}}{Super-population to use as reference panel. Default = "EUR". Options are EUR, SAS, EAS,
AFR, AMR. 'legacy' also available - which is a previously used verison of the EUR panel with a
slightly different set of markers}
}}

