% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_b_file}
\alias{get_b_file}
\title{get_b_file()}
\usage{
get_b_file(ancestry = "EUR")
}
\arguments{
\item{ancestry}{default is European 'EUR'}
}
\value{
the file path to the ld reference panel 'B' files
}
\description{
get_b_file()
}
