% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_class.R
\name{get_default_gwas_config}
\alias{get_default_gwas_config}
\title{get_default_gwas_config}
\usage{
get_default_gwas_config(open_file = FALSE)
}
\arguments{
\item{open_file}{whether to open file in R}
}
\value{
file path
}
\description{
opens the default config file and returns the path
}
