% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_downloads_dir}
\alias{get_downloads_dir}
\title{get_downloads_dir}
\usage{
get_downloads_dir(subdir = NA)
}
\arguments{
\item{subdir}{if provided with return the full file path to this subdirectory, within the main
downloads directory. If it does not exist, it will be created.}
}
\value{
the absoltue file path to the downloads folder
}
\description{
I need a downloads directory, just above the package directory, for development to stop package
building the large datasets. Hopefully will move this into inst/ext once the package is finished
}
