% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_finngen}
\alias{get_finngen}
\title{Gets the FinnGen data - downloads into inst/extdata}
\usage{
get_finngen(filename, type = "summary_stats", overwrite = FALSE)
}
\arguments{
\item{filename}{FinnGen file name, a string}

\item{type}{FinnGen data type}

\item{overwrite}{whether or not to overwrite the file}
}
\value{
a tibble containing the data
}
\description{
Gets the FinnGen data - downloads into inst/extdata
}
