% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proxy_snps.R
\name{get_proxy_snps}
\alias{get_proxy_snps}
\title{get_proxy_snps}
\usage{
get_proxy_snps(snps, ancestry = "EUR", ...)
}
\arguments{
\item{snps}{a character list or vector of SNPs}

\item{ancestry}{the ancestry to use default "EUR"}

\item{...}{additional paramters to pass to gwasvcf::get_ld_proxies}
}
\value{
a dataframe
}
\description{
get_proxy_snps
}
