% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_web_file}
\alias{get_web_file}
\title{get_web_file()}
\usage{
get_web_file(url, directory = NA, overwrite = FALSE)
}
\arguments{
\item{url}{url to file to download, a string}

\item{directory}{either NA (will download to main downloads directory), a full directory path
(will download here), or a string (will be used to create a subdirectory within the main
downloads directory)}

\item{overwrite}{whether or not to overwrite the file}
}
\value{
a file path to the data
}
\description{
Gets a web based file and downloads it into the packages downloads directory (inst/extdata)
}
