% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWebFile.R
\name{getWebFile}
\alias{getWebFile}
\title{Gets a web based file - downloads into inst/extdata}
\usage{
getWebFile(url, overwrite = FALSE)
}
\arguments{
\item{url}{url to file to download, a string}

\item{overwrite}{whether or not to overwrite the file}
}
\value{
a tibble containing the data
}
\description{
Gets a web based file - downloads into inst/extdata
}
