% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_class.R
\docType{class}
\name{GWAS-class}
\alias{GWAS-class}
\alias{GWAS}
\title{GWAS}
\value{
a GWAS object
}
\description{
This class handles GWAS data making it easy to import external GWASs, manipulate them and quickly
retrieve the results by loaded cached .csv files. It uses TwoSampleMR::format_data s
}
\section{Slots}{

\describe{
\item{\code{type}}{"exposure" or "outcome"}

\item{\code{file_path}}{path to a gwas file that can be read by data.table::fread()}

\item{\code{config}}{path to a yaml file with the column config information}

\item{\code{name}}{a name for this GWAS}

\item{\code{sig_pval}}{significance level at which to filter SNPs}

\item{\code{col_map}}{list of .}

\item{\code{overwrite}}{whether or not to overwrite the cached data files}

\item{\code{data}}{a data.table object with the parse GWAS SNP data}

\item{\code{gen_reg}}{a GenomeRegion object defining how to filter the SNP data}

\item{\code{clump_flag}}{a}
}}

