% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_class.R
\name{initialize,GWAS-method}
\alias{initialize,GWAS-method}
\title{initialize}
\usage{
\S4method{initialize}{GWAS}(.Object, ...)
}
\arguments{
\item{.Object}{a GWAS class object}

\item{...}{arguments passed during GWAS class instantiation}
}
\value{
a GWAS object
}
\description{
A custom initialize function / constructor for the GWAS class
}
