% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{setupMRLipidInfection}
\alias{setupMRLipidInfection}
\title{setupMRLipidInfection}
\usage{
setupMRLipidInfection(skip_local_download = TRUE)
}
\arguments{
\item{skip_local_download}{whether to skip the LD reference panel and PLINK downloads}
}
\value{
side effect setup functions (see below)
}
\description{
setupMRLipidInfection
}
