% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_class.R
\name{clump_snps,GWAS-method}
\alias{clump_snps,GWAS-method}
\title{Title}
\usage{
\S4method{clump_snps}{GWAS}(object, do_local = FALSE)
}
\arguments{
\item{object}{d}

\item{do_local}{d}
}
\value{
d
}
\description{
Title
}
