% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_class.R
\name{clump_snps,GWAS-method}
\alias{clump_snps,GWAS-method}
\title{Title}
\usage{
\S4method{clump_snps}{GWAS}(object, do_local = FALSE)
}
\arguments{
\item{do_local}{run on local machine, if false does on IEU servers}

\item{GWAS}{a GWAS object}
}
\value{
a GWAS object with clumped data
}
\description{
Title
}
