% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump_snps.R
\name{clump_snps}
\alias{clump_snps}
\title{clump_snps}
\usage{
clump_snps(object, do_local = FALSE)
}
\arguments{
\item{do_local}{whether to do the processing on the local computer or IEU servers}

\item{exposure_data}{a data.table consistent with output from TwoSampleMR::format_data; must be a
named list}

\item{genome_region}{a GenomeRegion object with appropriate clum}
}
\value{
clumped data.table, or list of them
}
\description{
clump_snps
}
