% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_significant_snps.R
\name{extract_significant_snps}
\alias{extract_significant_snps}
\title{extract_significant_snps}
\usage{
extract_significant_snps(
  filepath,
  name = NULL,
  pval = 5e-08,
  pval_col = "pvalue",
  snp_col = "SNP",
  overwrite = FALSE
)
}
\arguments{
\item{filepath}{a file path to extract from}

\item{name}{the name of the exposure / outcome}

\item{pval}{p-value filter (values less than this are kept)}

\item{pval_col}{the name of the column holding the p-value}

\item{snp_col}{the name of the column holding the snp id}

\item{overwrite}{whether or not to overwrite the cache .csv file}
}
\value{
the data table; and attribute file path to the filtered .csv file
}
\description{
extract_significant_snps
}
