% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_web_file}
\alias{get_web_file}
\title{Gets a web based file - downloads into inst/extdata}
\usage{
get_web_file(url, directory = NA, unzip = FALSE, overwrite = FALSE)
}
\arguments{
\item{url}{url to file to download, a string}

\item{directory}{either NA (will download to main downloads directory), a full directory path
(will download here), or a string (will be used to create a subdirectory within the main
downloads directory)}

\item{overwrite}{whether or not to overwrite the file}
}
\value{
a tibble containing the data
}
\description{
Gets a web based file - downloads into inst/extdata
}
