% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_class.R
\name{initialize,GWAS-method}
\alias{initialize,GWAS-method}
\title{initialize}
\usage{
\S4method{initialize}{GWAS}(.Object, ...)
}
\arguments{
\item{.Object}{d}

\item{...}{d}
}
\value{
a GWAS object
}
\description{
initialize
}
