% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{setupMRLipidInfection}
\alias{setupMRLipidInfection}
\title{setupMRLipidInfection}
\usage{
setupMRLipidInfection(skip_local_download = TRUE)
}
\arguments{
\item{skip_local_download}{whether to skip the LD reference panel and PLINK downloads}
}
\value{
side effect setup functions (see above)
}
\description{
Calling this function will download the 1000 genomes LD reference panel into this Package's
install directory. It will also ensure that PLINK is installed locally and if not it will
download it. The last thing that this does is ensure that the correct downloads directory exists
by calling \code{get_downloads_dir()}
}
